Use vec.push_back instead of vec.resize() to add timers with new ids.
Instead of externally managing the ids and use `resize()` if
necessary, we now distinguish how we add the items to the vector.

- If there are free ids from previous timers, then we re-use the id.
  This is preferred and works as before. In this case, we use
  `vec.insert(...)`
- If there is no free id, then we directly `vec.push_back(...)` into
  the vector instead of using resize + insert.

This is performs slightly better, since we no longer need to first create
and insert a default constructed Event, only to replace it with the correct
one later.