Use a size_t for the timer_id;
We should use `std::vector<T>::size_type`, because the timer_id is used
as index for the internal vector. But this means that the Event struct
needs to be available in the header. I'll think about making a header
only component, but for now we use size_t, which is correct for
the current case.