Use a std::set instead of a std::priority_queue to store the timeouts.
This has the advantage that when a timer event is removed, its id is
immediately available again. This is possible because the timeout
can be removed from the std::multiset, something which was not
(easily) possible with the priority_queue.

It has however the disadvantage that when a timer event is removed,
the std::multiset must be searched for the timers id in order to
remove the entry from the set. The search is linear because the set
is not sorted by ids (but by timeouts).

What is better depends on the use case. But I have found that this
behavior is preferable in many cases. One reason is the fact that the
number of entries in the vector are potentially smaller, because the
ids are immediately returned. The overhead of the search when an event
is removed is not a problem for small sets of timers.