Use a class to represent a timer, and move everything into cpptime.h.
This allows one to have multiple, independent timers. Another advantage
is that resources are only allocated once they are actually used.

Moving everyting into one header is more convenient to use the
component. Simply copy the header file.

One disadvantage of this change is that it is no longer possible to
manipulate (add, remove) timeouts through the functions in the CppTime
namespace. One needs a reference to a valid instance of a Timer.
However, to have the same effect, a client can create her own namespace
and wrap a CppTime::Timer there.