Cleanup Timer resources only after the run loop has finished.
Cleaning the resources (events) before the run loop has complete
can cause problems in the run loop when invoking the callback. The
reason is that at the time of invoking the callback, the events
are already cleared, which makes them invalid.

The correct behaviour is to wait until the run loop has complete before
removing all the events. This ensures that events are still valid at
time the callback is invoked.