Trigger immediate lambda cleanup on remove()
This change triggers an immediate lambda cleanup on remove(), so that any related resources can be cleaned up. 

I was getting a SEGV on exit in my program, because the CppTime Event object still existed at exit, causing delayed shared_ptr destruction, which in turn caused a crash with a dangling pointer. This change makes sure that all resources associated with a timer get cleaned up when it's removed.