Add tests for the handler cleanup for remove() and one-shot timeouts.
The last commit introduces a clean-up of the attached handler when
remove() is called or when a one-shot timer expires. This commit
adds two test cases to demonstrate the behaviour. A shared_ptr is
used for this, that will be cleaned-up (decrease the use_count) when
the handler is freed.

The commit also formats the code with clang-format (trivial change).